♻️ [refactor][#11] 회원의 송금 거래 이력 조회 엔드포인트 service단 빌더패턴으로 리팩토링
♻️[refactor][#11] 회원의 송금 거래 이력 조회 엔드포인트 service단 빌더패턴으로 리팩토링

+++ src/main/resources/table.sql
@@ -57,7 +57,7 @@ CREATE TABLE remittance_quote_v2 (
 CREATE TABLE remittance_log_v2 (
     log_id BIGINT AUTO_INCREMENT PRIMARY KEY,
     source_amount BIGINT NOT NULL,
-    fee double NOT NULL,
+    fee decimal(10, 0) NOT NULL,
     usd_exchange_rate decimal(10, 0) NOT NULL,
     usd_amount decimal(10, 0) NOT NULL,
     target_currency VARCHAR(255) NOT NULL,
